requirements.txt
streamlit>=1.12
pdfplumber>=0.7.6
python-docx>=0.8.11
requests>=2.28
openai>=1.0.0
python-dotenv>=0.21
pandas>=1.3





